crypto: msm: remove wakeup lock in qcrypto driver
When log server still runs and try to enable CE clock just before sleep,
wakeup lock will prevent the device sleep. And the bam/crypto is a not a
wakeupable device. As we have a state machine in crypto driver, and it
is able to control suspend behavior well, we do not need to use wakeup
lock in qcrypto driver.

Change-Id: I8824ac5ed66496edb1aaf5d6c1c9cd3e9f3ae7fc
Signed-off-by: Zhen Kong <zkong@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>